Built GET and POST network requests for flashcards
Built GET and POST network requests for flashcards, reorganized folders to create a dynamic route for flashcards (allowing for the GET and POST requests to access the slug as they are now in a dynamic route handler). Modified the prisma schema to remove unnecessary @@unique modifiers. Finished the signup route.ts file.