Studyguide UI + Studyguide backend
Created a page to view all studyguides with an option to create study guides at the top right and all the study guides listed ordered by created date. Created modal for creating study guides which calls a new post request to create study guides with tags. On the study guide list, if the study guide has a tag created box for tag to be displayed and when the study guide is clicked dynamic routing to /dashboard/[studyguideid], later showing the flashcards, videos, and practice questions for that study guide.