Added authentication and session checks
Implemented the NextAuth.js library and added Google OAuth2.0 authentication as well as checks on the server to show different webpages based on the user's login status. Likewise used authentication tokens to save the user's information on the postgresql database hosted on neon through the configuration for a prisma adapter with nextauth.