Created Baseline StudyGuide Menu
This has a map for all the study guides in the user (currently preset), used a ShadCN dropdown to sort the study guides, used a ShadCN Modal for a creation screen for studyguides and made a Create Tag mutli-select. There are also old components from an old rendition, they are only there to look back on later since it has been scrapped.