Request handler for studyguide and tag creation as well as minor UI changes
Added the POST request for creating a studyguide and the tag. Passed down a state so that whenever a study guide is created the useEffect for the filter automatically gets triggered, causing a rerender showing the new study guide. Added tag visibility on each of the study guides. Fixed a bug with creating tags via the dropdown menu (caused by a onBlur timeout being too quick resulting in erratic behavior). Reformatted tag placement in the studyguide display, as well as the studyguide display itself (ordered the divs by column instead of by row).