Completed all Network Requests
Created network requests for making/accessing/editing/deleting studysets, study method sets, and the study methods themselves. This is done by creating 3 more models in the schema to hold an array of videos/questions/flashcards. In each network request, the user is authenticated and a private check for data retrival is checked via the study guide.