Setup Skeleton and modified GET functionality
Created a basic skeleton ui for the study set page, then linked to the dynamic route holding the page from the /study page. Modified the GET function in /studyguides to allow for a query based on id, which would return a new custom prisma type, StudyGuide_f, containing both the tags and the studysets for the studyguide, as well as its actual data. Then used this information in the skeleton. However, the UI is still not complete and needs substantial reworking. Likewise, various functional features are necessary. For example, time spent on a studyguide or studyset is not tracked, and neither is the last accessed date. The actual stats do not show on click, and there is no modal for updating studyguides or creating new studysets. Likewise, there is no button to take one to the study set page. However, after these features are finished, this branch should be merged.