Completed Update Modal for editing studyguides
Repurposed modal for use in editing already created studyguides. Included checks to ensure that the name of the edited studyguide does not conflict with any others. As well as that, included checks to prevent users from continuously sending PATCH requests when no information has been changed. Lastly, added checkboxes to selectively update portions of the studyguide.