Created StudySetModal + Error Message Component + Reworked on Route
Reused StudyGuideModal (root) and made it into StudySetModal. Added an error message for those two modals specifically to show conflicts and add it to the bottom right.  Added background fade when modals are active, and on click (when there is no data thanks for pyro) to close the modal, and if there is data it won't close it.